retire Braintree & Scotland gis integrations
