+++ api.planx.uk/modules/gis/service/index.js
@@ -1,8 +1,6 @@
 import * as digitalLand from "./digitalLand.js";
-import * as scotland from "./local_authorities/scotland.js";
-import * as braintree from "./local_authorities/braintree.js";
 
-const localAuthorities = { digitalLand, braintree, scotland };
+const localAuthorities = { digitalLand };
 
 /**
  * @swagger
@@ -22,29 +20,12 @@ const localAuthorities = { digitalLand, braintree, scotland };
  *        description: Well-Known Text (WKT) formatted polygon or point
  */
 export async function locationSearch(req, res, next) {
-  // 'geom' param signals this localAuthority has data available via DigitalLand, "ready" teams are configured in PlanningConstraints component in editor
-  //   swagger doc intentionally doesn't cover legacy custom GIS hookups for Braintree and Scotland demo
+  // 'geom' param signals this localAuthority has data available via Planning Data
   if (req.query.geom) {
     try {
       const resp = await locationSearchWithoutTimeout(
         req.params.localAuthority,
         req.query,
-      );
-      res.send(resp);
-    } catch (err) {
-      next({
-        status: 500,
-        message: err ? err : "unknown error",
-      });
-    }
-    // check if this local authority is supported via our custom GIS hookup
-  } else if (localAuthorities[req.params.localAuthority]) {
-    try {
-      const timeout = Number(process.env.TIMEOUT_DURATION) || 15000;
-      const resp = await locationSearchWithTimeout(
-        req.params.localAuthority,
-        req.query,
-        timeout,
       );
       res.send(resp);
     } catch (err) {
@@ -61,37 +42,11 @@ export async function locationSearch(req, res, next) {
   }
 }
 
-// Digital Land is a single request with standardized geometry, so remove timeout & simplify query params
+// Planning Data is a single request with standardized geometry, so timeout is not necessary
 export function locationSearchWithoutTimeout(localAuthority, queryParams) {
   return localAuthorities["digitalLand"].locationSearch(
     localAuthority,
     queryParams.geom,
     queryParams,
   );
-}
-
-// custom GIS hookups require many requests to individual data layers which are more likely to timeout
-export function locationSearchWithTimeout(
-  localAuthority,
-  { x, y, siteBoundary, extras = "{}" },
-  time,
-) {
-  let extraInfo = extras;
-  extraInfo = JSON.parse(unescape(extras));
-
-  const timeout = new Promise((resolve, reject) => {
-    const timeoutID = setTimeout(() => {
-      clearTimeout(timeoutID);
-      reject("location search timeout");
-    }, time);
-  });
-
-  const promise = localAuthorities[localAuthority].locationSearch(
-    parseInt(x, 10),
-    parseInt(y, 10),
-    JSON.parse(siteBoundary),
-    extraInfo,
-  );
-
-  return Promise.race([promise, timeout]);
 }
