+++ api.planx.uk/modules/gis/service/index.test.ts
@@ -2,72 +2,6 @@ import supertest from "supertest";
 
 import loadOrRecordNockRequests from "../../../tests/loadOrRecordNockRequests.js";
 import app from "../../../server.js";
-import { locationSearchWithTimeout } from "./index.js";
-
-// Tests commented out due to reliance on external API calls and fallibility of nocks
-//   Please comment in and run locally if making changes to /gis functionality
-describe("locationSearchWithTimeout", () => {
-  beforeEach(() => {
-    vi.useFakeTimers();
-  });
-
-  afterEach(() => {
-    vi.useRealTimers();
-  });
-
-  test.skip("a successful call", async () => {
-    const timeout = 500;
-    const localAuthority = "braintree";
-    const promise = locationSearchWithTimeout(
-      localAuthority,
-      { x: 50, y: 50, siteBoundary: "[]" },
-      timeout,
-    );
-    await expect(promise).resolves.toStrictEqual(expect.any(Object));
-  });
-
-  test.skip("an immediate timeout", async () => {
-    const timeout = 500;
-    const localAuthority = "braintree";
-    const promise = locationSearchWithTimeout(
-      localAuthority,
-      { x: 50, y: 50, siteBoundary: "[]" },
-      timeout,
-    );
-    vi.runAllTimers();
-    await expect(promise).rejects.toEqual("location search timeout");
-  });
-});
-
-describe.skip("fetching GIS data from local authorities directly", () => {
-  const locations = [
-    {
-      council: "braintree",
-      x: 575629.54,
-      y: 223122.85,
-      siteBoundary: [],
-    },
-  ];
-
-  loadOrRecordNockRequests("fetching-direct-gis-data", locations);
-
-  locations.forEach((location) => {
-    it(`returns MVP planning constraints for ${location.council}`, async () => {
-      await supertest(app)
-        .get(
-          `/gis/${location.council}?x=${location.x}&y=${
-            location.y
-          }&siteBoundary=${JSON.stringify(location.siteBoundary)}`,
-        )
-        .expect(200)
-        .then((res) => {
-          expect(res.body["article4"]).toBeDefined();
-          expect(res.body["listed"]).toBeDefined();
-          expect(res.body["designated.conservationArea"]).toBeDefined();
-        });
-    }, 20_000); // 20s request timeout
-  });
-});
 
 describe.skip("fetching GIS data from Digital Land for supported local authorities", () => {
   const locations = [
